changes to frmMain and Agents
Included notepad file with the code to be added to Package.DB,
Supplier.DB and Product.Db. I didn't want to upload my copy of these
files to GitHub just in case :) Also added some notes/questions to the
code for some things i found. Lots to do for the final project but its
looking great!!!